reorganized gui files
