--- ClientGUI/packages.config
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="PInvoke.Kernel32" version="0.7.124" targetFramework="net451" />
-  <package id="PInvoke.SetupApi" version="0.7.124" targetFramework="net451" />
-  <package id="PInvoke.Windows.Core" version="0.7.124" targetFramework="net451" />
-  <package id="System.Buffers" version="4.5.1" targetFramework="net451" />
-  <package id="System.Memory" version="4.5.4" targetFramework="net451" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.3" targetFramework="net451" />
-  <package id="System.Threading.Tasks.Extensions" version="4.5.4" targetFramework="net451" />
-  <package id="System.ValueTuple" version="4.5.0" targetFramework="net451" />
-</packages>
